backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
httpx==0.24.1
python-dotenv==1.0.0
asgiref==3.8.1
